Rename gc module
